requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.8.2  # Updated from 2.4.2
httpx==0.25.0
langchain
langchain_core
